requirements.txt
absl-py==2.2.2
accelerate==1.6.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
contourpy==1.3.1
cycler==0.12.1
diffusers==0.32.2
filelock==3.13.1
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.6.1
grpcio==1.71.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.8
lightning==2.5.1
lightning-utilities==0.14.3
lpips==0.1.4
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.1
mpmath==1.3.0
multidict==6.3.2
networkx==3.3
numpy==1.26.1
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.1.105
nvidia-nvtx-cu12==12.1.105
opencv-python==4.11.0.86
pandas==2.2.3
pillow==11.0.0
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
pyparsing==3.2.3
pytorch-fid==0.3.0
pytorch-lightning==2.5.1
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scipy==1.15.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
timm==1.0.15
torch==2.2.1+cu121
torchaudio==2.2.1+cu121
torchmetrics==1.7.0
torchvision==0.17.1+cu121
tqdm==4.67.1
triton==2.2.0
tzdata==2025.2
urllib3==2.3.0
Werkzeug==3.1.3
yarl==1.19.0
